requirements.txt
pyserial==3.2.1
Flask==0.12
